fastapi-auth/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
httpx
python-jose[cryptography]
pydantic[email]
alembic
python-dotenv
